+++ test/eslint-plugin-e2e/test/configs/monorepo.config.js
@@ -22,6 +22,7 @@ export default [
           capture: ["name"],
         },
       ],
+      "boundaries/dependency-nodes": ["import"],
     },
     rules: {
       ...strictBoundariesConfig.rules,